Preliminary base for service tool
At this point it can load a .dbc file (it is hard-coded to use a signal from AFE_CAN_ID247_FACTORY.dbc) and subsequently display a GUI with a single progress bar being driven by CAN messages of that signal.  In Linux it is presently using the vcan0 interface.

Libraries:
* Python3.4
* PyQt5
* python-can
* canmatrix

Ref T105

Cc: tsnider