Use .ui with properties referencing CAN signals
Introduced a QtDesigner file and the ability to automatically
connect CAN signals based the the 'frame' and 'signal' properties
of the Qt widgets added to the Designer file.  These properties
can be added dynamically to any widget so there is no need to
subclass all existing widgets to get this 'auto-connect' feature.

Ref T105