Fix silent crash when deployed EXE output is not redirected to file
This was introduced when initially getting pyqtdeploy working [1] because I added sys.stdout.Flush() as a diagnostic attempt when no messages were printed to the screen.  The result at first was that when the application was deployed as a GUI, sys.stdout was None and hence an exception was thrown.  Without any stdout, the exception had nowhere to be reported.  Later I added the GUI exception dialog but sys.stdout.flush() was called before I created my QApplication() so another exception was thrown when I attempted to launch the QMessageBox() to report the first exception.  But, it seems to all be unwound now.

[1] https://epc-phab.exana.io/rST6c838141d1931abf43301bfcf2d8ac8abc4581d4#inline-247

Ref T168