Add device.ui wrapper around individual devices ui
This provides the old Tx/Rx and NV tabs along with a dash tab that shows the individual devices dash.  I would expect that in the future the device would be able to configure some of these 'default' behaviors'

Note that this commit is not appropriately attaching the Tx/Rx and NV tabs to the bus.  Only the GUI change is implemented here.

Ref T177