Add script to build EPyQ installer
It's not really a 'build system' but at least you don't have to remember
all the magical CLI and GUI incantations and hope that it finishes
within 99 seconds...

Includes venv update, revision hash file generation, and locates the
installer in the root with a name including the present revision hash.

Log available in root as `build.log`.

Subscribers: tsnider

Ref T197