Agree on `bits_per_byte`, fix Chunk.update() slicing
From c99 limits.h CHAR_BIT is "number of bits for smallest object
that is not a bit-field (byte)" and shall have a value >= 8.  So
having a uint16_t as the smallest type and 16-bits per address and
reporting a uint16_t as being a single byte might actually be
conformant.

http://stackoverflow.com/a/5516161/228539